delete _importlib_modulespec
